requirements.txt
streamlit
roboticstoolbox-python
pandas
numpy
watchdog
matplotlib
spatialmath-python